requirements.txt
# ============================================================
# AI-Based Exam Anxiety Detector - Requirements
# ============================================================

# Core ML & NLP
torch==2.2.0
transformers==4.35.2
datasets==2.15.0
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.3

# Model Training & Evaluation
accelerate==0.24.1
evaluate==0.4.1
seqeval==1.2.2

# Backend API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2

# Frontend
streamlit==1.28.2
plotly==5.18.0
altair==5.2.0

# Data Processing & Visualization
matplotlib==3.8.2
seaborn==0.13.0
wordcloud==1.9.3
nltk==3.8.1
regex==2023.10.3

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tqdm==4.66.1
joblib==1.3.2

# Notebook support
jupyter==1.0.0
ipywidgets==8.1.1
